requirements.txt
Flask
python-dotenv
cryptography
requests
# optional: flask-limiter for rate limiting
flask-limiter
